Totally changed models. Change db from local to server deployed

+++ AgileDiary2/Models/Goal.cs
@@ -4,11 +4,11 @@ namespace AgileDiary2.Models
 {
     public class Goal
     {
-        public ulong GoalId { get; set; }
+        public int GoalId { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public string Reward { get; set; }
-        public ulong SprintId { get; set; }
+        public int SprintId { get; set; }
         public ICollection<MyTask> Milestones { get; set; }
         public Area Area { get; set; }
     }
